Create Dog class definition
Write a class definition line and a one line docstring for the class Dog. Write an __init__ method for the class Dog that gives each dog its own name and breed. Test this on a successful creation of a Dog object.
>>> import dog
>>> sugar = dog.Dog('Sugar', 'border collie')
>>> sugar.name
Sugar
>>> sugar.breed
border collie
Problem 2
Add a data attribute tricks of type list to each Dog instance and initialize it in __init__ to the empty list. The user does not have to supply a list of tricks when constructing a Dog instance. Make sure that you test this successfully.
>>> sugar.tricks
[]
Problem 3
Write a method teach as part of the class Dog. The method teach should add a passed string parameter to tricks and print a message that the dog knows the trick.
>>> sugar.teach('frisbee')
Sugar knows frisbee